refactor: extract JS helpers from postbuild-fixup into lib/helpers.js
Phase 1 of #487: Moves ~270 lines of string-appended JS code into a
proper source file (lib/helpers.js). The postbuild script now appends a
10-line require() bridge instead of string-concatenating helper
functions. This makes the helpers testable, lintable, and editable as
normal source code.

Closes #487
