+++ package.json
@@ -95,6 +95,7 @@
     "index.js",
     "index.mjs",
     "index.d.ts",
+    "lib/",
     "streaming/",
     "README.md",
     "README.ja.md",
